Support VTK's autoinit with a different approach
The recipe generates the required autoinit header files, and uses
special definitions to enable the appropriate parts of the autoinit
files.

This is a different approach than what VTK uses, and should work with
all build systems.

There are likely other things that need to be supported, but one thing
at a time ...