cpython: always build a shared lib for the interpreter, tighten env vars
Using a static library for the interpreter and modules is very fragile and does not support the modules well.
Always build a working interpreter with all modules, even for shared=False, by linking the interpreter against a shared version.
This is necessary to ensure that consumers can deploy a working interpreter and modules with the correct host architecture regardless of shared=True/False.